social_network/requirements.txt
asgiref==3.11.1
Django==5.0.2
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
pytz==2026.1.post1
sqlparse==0.5.5
